### 2025-04-18 #8081 @editorialbot
[PRE REVIEW]: Polyatomic Complexes: A Software Framework for Topologically Accurate Representations of Molecules
**Submitting author:** @rahulkhorana (<a href="http://orcid.org/0000-0001-8795-1623">Rahul Khorana</a>)
**Repository:** https://github.com/rahulkhorana/PolyatomicComplexes
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.8
