### 2025-04-18 89474a5
Bump torch in /requirements in the pip group across 1 directory
Bumps the pip group with 1 update in the /requirements directory: [torch](https://github.com/pytorch/pytorch).


Updates `torch` from 2.5.1 to 2.6.0
- [Release notes](https://github.com/pytorch/pytorch/releases)
- [Changelog](https://github.com/pytorch/pytorch/blob/main/RELEASE.md)
- [Commits](https://github.com/pytorch/pytorch/compare/v2.5.1...v2.6.0)

---
updated-dependencies:
- dependency-name: torch
  dependency-version: 2.6.0
  dependency-type: direct:production
  dependency-group: pip
...

Signed-off-by: dependabot[bot] <sup
[…]
