+++ requirements/requirements_graph.txt
@@ -187,7 +187,7 @@ terminado==0.18.1
 threadpoolctl==3.5.0
 tinycss2==1.4.0
 TopoNetX==0.2.0
-torch==2.5.1
+torch==2.6.0
 tornado==6.4.2
 tqdm==4.67.1
 traitlets==5.14.3
