+++ paper.bib
@@ -5,6 +5,7 @@
       eprint={2409.15600},
       archivePrefix={arXiv},
       primaryClass={cs.LG},
+      doi = {10.48550/arXiv.2409.15600},
       url={https://arxiv.org/abs/2409.15600}, 
 }
 
@@ -16,6 +17,7 @@
   number={1},
   pages={41},
   year={2022},
+  doi = {10.1038/s41524-022-00721-x},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -39,6 +41,7 @@
       year={2017},
       eprint={1703.02833},
       archivePrefix={arXiv},
+      doi = {10.48550/arXiv.1703.02833},
       primaryClass={quant-ph},
       url={https://arxiv.org/abs/1703.02833}, 
 }
@@ -50,6 +53,7 @@
   volume={153},
   number={2},
   year={2020},
+  doi = {10.1063/5.0006074},
   publisher={AIP Publishing}
 }
 
@@ -60,6 +64,7 @@
   volume={68},
   pages={314--319},
   year={2013},
+  doi = {https://doi.org/10.1016/j.commatsci.2012.10.028},
   publisher={Elsevier}
 }
 
@@ -71,6 +76,7 @@
   number={4},
   pages={77},
   year={2024},
+  doi = {10.48550/arXiv.2302.03836},
   publisher={Springer}
 }
 
@@ -81,6 +87,7 @@
       eprint={2408.16686},
       archivePrefix={arXiv},
       primaryClass={cs.LG},
+      doi = {10.48550/arXiv.2408.16686},
       url={https://arxiv.org/abs/2408.16686}, 
 }
 
@@ -93,6 +100,7 @@
   number={1},
   pages={31--36},
   year={1988},
+  doi={10.1021/ci00057a005},
   publisher={ACS Publications}
 }
 
@@ -102,6 +110,7 @@
   journal={Handbook of Materials Modeling: Methods: Theory and Modeling},
   pages={1751--1784},
   year={2020},
+  doi={10.1007/978-3-319-42913-7_60-1},
   publisher={Springer}
 }
 
@@ -111,6 +120,7 @@
   journal={The Journal of chemical physics},
   volume={96},
   number={10},
+  doi = {10.1063/1.462413},
   pages={7603--7614},
   year={1992},
   publisher={American Institute of Physics}
@@ -124,6 +134,7 @@
   number={5},
   pages={742--754},
   year={2010},
+  doi={10.1021/ci100050t},
   publisher={ACS Publications}
 }
 
@@ -134,6 +145,7 @@
   volume={3},
   number={10},
   year={2022},
+  doi = {https://doi.org/10.1016/j.patter.2022.100588},
   publisher={Elsevier}
 }
 
@@ -141,7 +153,8 @@
   title={Gaussian Processes on Cellular Complexes},
   author={Alain, Mathieu and Takao, So and Paige, Brooks and Deisenroth, Marc Peter},
   booktitle={International Conference on Machine Learning},
-  year={2024}
+  year={2024},
+  doi = {10.48550/arXiv.2311.01198},
 }
 
 @article{BhadwalGenSMILES2023,
@@ -151,7 +164,7 @@ volume = {268},
 pages = {110429},
 year = {2023},
 issn = {0950-7051},
-doi = {https://doi.org/10.1016/j.knosys.2023.110429},
+doi = {10.1016/j.knosys.2023.110429},
 url = {https://www.sciencedirect.com/science/article/pii/S095070512300179X},
 author = {Arun Singh Bhadwal and Kamal Kumar and Neeraj Kumar},
[…]
