+++ paper.bib
@@ -64,7 +64,7 @@
   volume={68},
   pages={314--319},
   year={2013},
-  doi = {https://doi.org/10.1016/j.commatsci.2012.10.028},
+  doi = {10.1016/j.commatsci.2012.10.028},
   publisher={Elsevier}
 }
 
@@ -145,7 +145,7 @@
   volume={3},
   number={10},
   year={2022},
-  doi = {https://doi.org/10.1016/j.patter.2022.100588},
+  doi = {10.1016/j.patter.2022.100588},
   publisher={Elsevier}
 }
 
