+++ paper.md
@@ -45,9 +45,9 @@ The software is helpful to researchers in cheminformatics, quantum chemistry, an
 
 While widely used, current molecular representations—such as SMILES, SELFIES, molecular graphs, and ECFP fingerprints—often fail to incorporate physically meaningful topological and electronic structure information [@khorana2024polyatomiccomplexestopologicallyinformedlearning] [@manolopoulos1992molecular] [@rogers2010extended] [@krenn2022selfies]. These representations, although computationally efficient, are not designed to satisfy key scientific constraints such as topological accuracy, long-range interactions, and differentiability with respect to atomic coordinates [@langer2022representations] [@BhadwalGenSMILES2023] [@le2020neuraldecipher].
 
-Polyatomic Complexes address this gap by introducing a general-purpose, extensible, and topologically rigorous representation that integrates smoothly with modern ML pipelines. This fills a critical unmet need in the intersection of chemistry, materials science, and machine learning.
+Polyatomic Complexes address this gap by introducing a general-purpose, and topologically informed representation that integrates smoothly with modern ML pipelines. This fills a critical unmet need in the intersection of chemistry, materials science, and machine learning.
 
-Moreover with the advent of topological deep learning and geometric deep learning such representations will become increasingly applicable [@zia2024topological]. A classic example of this is Cellular Neural Networks [@khorana2024cwcnncwan] or Cellular Gaussian Processes [@alain2024gaussian] which are compatible with Polyatomic Complexes [@khorana2024polyatomiccomplexestopologicallyinformedlearning].
+Moreover with the advent of topological deep learning such representations will become increasingly applicable [@zia2024topological]. A classic example of this is Cellular Neural Networks [@khorana2024cwcnncwan] or Cellular Gaussian Processes [@alain2024gaussian] which are compatible with Polyatomic Complexes [@khorana2024polyatomiccomplexestopologicallyinformedlearning].
 
 # Code Contributions and Workflow
 
@@ -56,7 +56,7 @@ The provided code enables researchers to develop effective models for a variety 
 1. An implementation of the core Polyatomic Complexes representation
 2. An easy to use, well documented API for experiments, see the [official documentation](https://rahulkhorana.github.io/PolyatomicComplexes/).
 3. Easy integration with existing quantum chemistry libraries such as pyscf [@sun2020recent], and pymatgen [@ong2013python].
-4. Nicely packaged example datasets for baseline performance and benchmarking.
+4. Well packaged example datasets for baseline performance and benchmarking.
 
 Moreover the Polyatomic Complexes are stratified into different categories depending on usage and experimental need. This is touched upon in the Software Description.
 
@@ -64,7 +64,7 @@ A standard workflow for molecular machine learning using Polyatomic Complexes wo
 
 ![In this figure we see a standard molecular ML pipeline including Polyatomic Complexes.](figures/figure.png)
 
[…]
