### 2025-04-21 06281a3
Bump numpy from 1.26.4 to 2.0.2
Bumps [numpy](https://github.com/numpy/numpy) from 1.26.4 to 2.0.2.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/main/doc/RELEASE_WALKTHROUGH.rst)
- [Commits](https://github.com/numpy/numpy/compare/v1.26.4...v2.0.2)

---
updated-dependencies:
- dependency-name: numpy
  dependency-version: 2.0.2
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
