+++ requirements/requirements_graph.txt
@@ -115,7 +115,7 @@ nbformat==5.10.4
 nest-asyncio==1.6.0
 networkx==3.4.2
 notebook_shim==0.2.4
-numpy==1.26.4
+numpy==2.0.2
 opt_einsum==3.4.0
 ordered-set==4.1.0
 orderly-set==5.2.3
