### 2025-05-08 #8081 @ritesh001
> Dear [@JacksonBurns](https://github.com/JacksonBurns), [@WangKehan573](https://github.com/WangKehan573), [@leeberty84](https://github.com/leeberty84), [@obliviateandsurrender](https://github.com/obliviateandsurrender), [@peterhys](https://github.com/peterhys), [@ritesh001](https://github.com/ritesh001),
>
> I am currently in the review pipeline at JOSS and would like to ask if any of you might be willing to serve as a reviewer.
>
> I completely understand if you are unavailable or if it is not a good time. Thank you very much for considering this request!

Apologies for the late reply. I am willing to be a reviewer.
