### 2025-05-12 #8081 @RMeli
Hi @rahulkhorana, thank you for the ping. I still need to evaluate the code myself, before going forward.

For example, I noticed the presence of a lot of duplicated code across the main classes of the package in `src/experiment`. The different `Pricess*` classes all seem to have the functions with the same code apart from a few different paths. I also started to notice a lot of code duplication elsewhere (see for example the `Proton` and `Neutron` class), therefore I will need more time to evaluate if this submission can be moved to the review stage.

Thank you for your patience.
