### 2025-05-12 #8081 @rahulkhorana
Hi @RMeli,

Thank you for the update, and I completely understand. Please take all the time you need to properly evaluate the code.

I appreciate you pointing out the areas of duplication — that’s very helpful feedback. I’ll make a note of the issues you mentioned, particularly in the Process, Proton, and Neutron classes, and will be happy to refactor those as needed based on your final assessment.

Looking forward to hearing more once you’ve had a chance to complete the review.

Best regards,
Rahul
