### 2025-05-12 #8081 @rahulkhorana
Hi @RMeli,

Thank you again for the feedback — it was really helpful in highlighting areas for improvement.

I’ve now pushed a set of changes that refactor the duplicated logic across the `Process*` classes in src/experiment. These are now all unified via a shared `BaseProcess` class. I’ve also cleaned up the `Proton` and `Neutron` classes by abstracting their shared logic into a common `Particle` base class.

The changes are reflected in master per commit: `287ed97.`

Please let me know if you’d like any specific parts highlighted or if there’s anything else I should address before review.

Thanks again for your guidance,
