### 2025-06-30 #8081 @rahulkhorana
Hi @RMeli, I'm doing well. Thank you for asking. The refactoring is mostly done. The only thing that remains to be refactored is the experiment folders. I think that folder will end up being deprecated/removed in a later release. I can refactor that folder this week. Otherwise we should be good to go!
