### 2025-08-06 rahulkhorana/PolyatomicComplexes#157 @rahulkhorana
Duplication and Data
**Describe the bug**

https://github.com/openjournals/joss-reviews/issues/8081#issuecomment-3159912702

For example, experiments have a long if/else pattern where many entries appear to be the same. This is repeated across the different experiments.

https://github.com/rahulkhorana/PolyatomicComplexes/blob/2091e7f42fc496a18f43bc6c4a221a013d35c39e/polyatomic_complexes/experiments/freesolv_experiment.py#L136-L184
https://github.com/rahulkhorana/PolyatomicComplexes/blob/2091e7f42fc496a18f43bc6c4a221a013d35c39e/polyatomic_complexes/experiments/lipophilicity_experiment.py#L141-L189
In my opinion, the polyatomic_complexes/results folder does not belong to repository, since it contains a lot of images. Likewise, it is debatable if the data should be stored alongside the code or would be better suited elsewhere.
