+++ polyatomic_complexes/experiments/freesolv_experiment.py
@@ -1,291 +1,0 @@
-import os
-import torch
-import time
-import numpy as np
-from .load_process_data import LoadDatasetForTask
-
-# botorch specific
-from botorch.models.gp_regression import ExactGP
-
-# gpytorch specific
-from gpytorch.means import ConstantMean
-from gpytorch.kernels import ScaleKernel, RBFKernel, Kernel
-from gpytorch.distributions import MultivariateNormal
-from gpytorch.likelihoods import GaussianLikelihood, Likelihood
-from gpytorch.mlls import ExactMarginalLogLikelihood
-
-# kernels + gp
-from .kernels import TanimotoKernel
-from .gaussian_process import evaluate_model, evaluate_graph_model
-from gauche import SIGP, NonTensorialInputs
-from gauche.kernels.graph_kernels import (
-    ShortestPathKernel,
-    VertexHistogramKernel,
-    GraphletSamplingKernel,
-)
-
-from matplotlib import pyplot as plt
-
-if torch.cuda.is_available():
-    dev = "cuda:0"
-    torch.cuda.empty_cache()
-else:
-    dev = "cpu"
-device = torch.device(dev)
-
-
-class ExactGPModel(ExactGP):
-    def __init__(self, train_x, train_y, likelihood):
-        super(ExactGPModel, self).__init__(train_x, train_y, likelihood)
-        self.mean_module = ConstantMean()
-        self.covar_module = ScaleKernel(TanimotoKernel())
-
-    def forward(self, x):
-        mean_x = self.mean_module(x)
-        covar_x = self.covar_module(x)
-        return MultivariateNormal(mean_x, covar_x)
-
-
-class GraphGP(SIGP):
-    def __init__(self, train_x, train_y, likelihood, kernel, **kernel_kwargs):
-        super().__init__(train_x, train_y, likelihood)
-        self.mean = ConstantMean()
-        self.covariance = kernel(**kernel_kwargs)
-
-    def forward(self, x):
-        """
-        A forward pass through the model.
-        """
-        mean = self.mean(torch.zeros(len(x), 1)).float()
-        covariance = self.covariance(x)
-
-        # because graph kernels operate over discrete inputs it is beneficial
-        # to add some jitter for numerical stability
-        jitter = max(covariance.diag().mean().detach().item() * 1e-4, 1e-4)
-        covariance += torch.eye(len(x)) * jitter
-        return MultivariateNormal(mean, covariance)
-
-
-class StackedGP(SIGP):
-    def __init__(
-        self,
-        train_x: NonTensorialInputs,
-        train_y: torch.Tensor,
-        likelihood: Likelihood,
-        kernel: Kernel,
-    ):
-        super().__init__(train_x, train_y, likelihood)
-        self.mean = ConstantMean()
-        self.covariance = kernel()
-
-    def forward(self, x):
-        """
-        A forward pass through the model.
-        """
-        mean = self.mean(torch.zeros(len(x), 1)).float()
-        covariance = self.covariance(x)
-
-        # because graph kernels operate over discrete inputs it is beneficial
-        # to add some jitter for numerical stability
-        jitter = max(covariance.diag().mean().detach().item() * 1e-4, 1e-4)
-        covariance += torch.eye(len(x)) * jitter
-        return MultivariateNormal(mean, covariance)
-
-
-def initialize_model(train_x: torch.Tensor, train_obj: torch.Tensor, likelihood):
-    model = ExactGPModel(train_x, train_obj, likelihood).to(train_x)
-    return model
-
-
-def initialize_graph_gp(train_x, train_obj, likelihood, kernel, **kernel_kwargs):
[…]
