+++ polyatomic_complexes/experiments/__init__.py
@@ -1,11 +1,6 @@
-from .esol_experiment import *
-from .freesolv_experiment import *
-from .gaussian_process import *
-from .jdft2d_matbench_experiment import *
 from .kernels import *
-from .lipophilicity_experiment import *
-from .load_process_data import *
-from .materials_project_experiment import *
 from .metrics import *
-from .mp_matbench_experiment import *
-from .photoswitches_experiment import *
+from .run_experiment import *
+from .gaussian_process import *
+from .load_process_data import *
+from .experiment_runner import *

[…]
