+++ polyatomic_complexes/results/Lipophilicity/fast_complexes_Results/complexes_1712114802.721913.txt
@@ -1,5 +1,0 @@
-Lipophilicity:
-complexes:column: exp, 
-mean R^2: -0.0006 +- 0.0002, mean RMSE: 1.2073 +- 0.0050, mean MAE: 0.9773 +- 0.0041
-, mean CRPS: 0.9773 +- 0.0041
-
