+++ polyatomic_complexes/results/ESOL/deep_complexes_Results/deep_complexes_1712113774.30992.txt
@@ -1,33 +1,0 @@
-ESOL:
-deep_complexes:column: ESOL predicted log solubility in mols per litre, 
-mean R^2: -0.0059 +- 0.0021, mean RMSE: 1.6763 +- 0.0109, mean MAE: 1.3166 +- 0.0083
-, mean CRPS: 1.3166 +- 0.0083
-
-column: Minimum Degree, 
-mean R^2: -0.0044 +- 0.0019, mean RMSE: 0.2358 +- 0.0043, mean MAE: 0.1111 +- 0.0011
-, mean CRPS: 0.1111 +- 0.0011
-
-column: Molecular Weight, 
-mean R^2: -0.0031 +- 0.0008, mean RMSE: 102.0077 +- 0.6070, mean MAE: 83.5198 +- 0.4137
-, mean CRPS: 83.5198 +- 0.4137
-
-column: Number of H-Bond Donors, 
-mean R^2: -0.0041 +- 0.0015, mean RMSE: 1.1126 +- 0.0173, mean MAE: 0.7991 +- 0.0065
-, mean CRPS: 0.7991 +- 0.0065
-
-column: Number of Rings, 
-mean R^2: -0.0041 +- 0.0011, mean RMSE: 1.3184 +- 0.0106, mean MAE: 1.0529 +- 0.0057
-, mean CRPS: 1.0529 +- 0.0057
-
-column: Number of Rotatable Bonds, 
-mean R^2: -0.0038 +- 0.0008, mean RMSE: 2.6571 +- 0.0441, mean MAE: 1.9449 +- 0.0172
-, mean CRPS: 1.9449 +- 0.0172
-
-column: Polar Surface Area, 
-mean R^2: -0.0051 +- 0.0018, mean RMSE: 35.9267 +- 0.3652, mean MAE: 28.2431 +- 0.2187
-, mean CRPS: 28.2431 +- 0.2187
-
-column: measured log solubility in mols per litre, 
-mean R^2: -0.0045 +- 0.0021, mean RMSE: 2.0860 +- 0.0137, mean MAE: 1.6541 +- 0.0102
-, mean CRPS: 1.6541 +- 0.0102
-
