+++ polyatomic_complexes/results/ESOL/SMILES_Results/SMILES_1712114188.379835.txt
@@ -1,33 +1,0 @@
-ESOL:
-SMILES:column: ESOL predicted log solubility in mols per litre, 
-mean R^2: 0.8882 +- 0.0035, mean RMSE: 0.5581 +- 0.0104, mean MAE: 0.3944 +- 0.0043
-, mean CRPS: 0.9513 +- 0.0121
-
-column: Minimum Degree, 
-mean R^2: 0.5830 +- 0.0086, mean RMSE: 0.1517 +- 0.0031, mean MAE: 0.0714 +- 0.0012
-, mean CRPS: 0.0651 +- 0.0022
-
-column: Molecular Weight, 
-mean R^2: 0.9060 +- 0.0046, mean RMSE: 31.0630 +- 0.8469, mean MAE: 18.8214 +- 0.2970
-, mean CRPS: 56.3289 +- 0.6658
-
-column: Number of H-Bond Donors, 
-mean R^2: 0.8374 +- 0.0055, mean RMSE: 0.4457 +- 0.0085, mean MAE: 0.3083 +- 0.0028
-, mean CRPS: 0.5350 +- 0.0073
-
-column: Number of Rings, 
-mean R^2: 0.8806 +- 0.0040, mean RMSE: 0.4540 +- 0.0101, mean MAE: 0.2483 +- 0.0030
-, mean CRPS: 0.7052 +- 0.0110
-
-column: Number of Rotatable Bonds, 
-mean R^2: 0.8535 +- 0.0063, mean RMSE: 1.0119 +- 0.0309, mean MAE: 0.6127 +- 0.0071
-, mean CRPS: 1.3559 +- 0.0203
-
-column: Polar Surface Area, 
-mean R^2: 0.8830 +- 0.0044, mean RMSE: 12.2290 +- 0.2991, mean MAE: 7.9743 +- 0.1181
-, mean CRPS: 19.3635 +- 0.2210
-
-column: measured log solubility in mols per litre, 
-mean R^2: 0.8427 +- 0.0028, mean RMSE: 0.8242 +- 0.0061, mean MAE: 0.6385 +- 0.0045
-, mean CRPS: 1.2046 +- 0.0152
-
