+++ polyatomic_complexes/results/FreeSolv/SMILES_Results/SMILES_1712114664.393806.txt
@@ -1,9 +1,0 @@
-FreeSolv:
-SMILES:column: expt, 
-mean R^2: 0.7642 +- 0.0105, mean RMSE: 1.8712 +- 0.0468, mean MAE: 1.3883 +- 0.0228
-, mean CRPS: 2.0679 +- 0.0316
-
-column: calc, 
-mean R^2: 0.8147 +- 0.0046, mean RMSE: 1.8042 +- 0.0233, mean MAE: 1.3399 +- 0.0150
-, mean CRPS: 2.2724 +- 0.0308
-
