### 2025-08-07 105a89c
closes issue #157
