### 2025-08-07 #8081 @rahulkhorana
@RMeli

The issues you mentioned have been resolved with the closing of https://github.com/rahulkhorana/PolyatomicComplexes/issues/157#issuecomment-3164338305.

**Summary of changes:**

- The experiment has been refactored.
- The results folder has been removed.

All changes are reflected in the repo on branch `joss-submission`.

Thanks!
Rahul
