### 2025-08-13 #8081 @RMeli
Thanks @rahulkhorana. I see that the changes have been made to the `joss` branch. Any chances you can move them to the `master` branch? Otherwise the reviewers might miss some things, since the `master` branch is the default.
