+++ README.md
@@ -12,7 +12,7 @@
   <img src="https://img.shields.io/pypi/v/polyatomic-complexes?style=plastic&logo=%233775A9&logoSize=auto&labelColor=%233775A9&color=%23e1ad01&link=https%3A%2F%2Fpypi.org%2Fproject%2Fpolyatomic-complexes%2F0.0.8%2F" alt="PyPI Version">
   <img src="https://img.shields.io/pypi/format/polyatomic-complexes" alt="PyPI Format">
   <a href="https://pepy.tech/project/polyatomic-complexes">
-  <img src="https://img.shields.io/endpoint?url=https://pepy.tech/api/projects/polyatomic-complexes&style=plastic&color=e1ad01&label=downloads" alt="Downloads">
+  <img src="https://pepy.tech/project/polyatomic-complexes&style=plastic&color=e1ad01&label=downloads" alt="Downloads">
 </a>
   <a href="https://socket.dev/pypi/package/polyatomic-complexes/overview/1.0.7/tar-gz">
     <img src="https://socket.dev/api/badge/pypi/package/polyatomic-complexes/1.0.7?artifact_id=tar-gz" alt="Socket Badge">
