### 2025-08-18 02761a9
JOSS: paper-only branch (no LFS)
