+++ polyatomic_complexes/__init__.py
@@ -1,2 +1,0 @@
-from .src import *
-from .experiments import *
