+++ polyatomic_complexes/src/__init__.py
@@ -1,2 +1,0 @@
-from .complexes import *
-from .experiment import *

[…]
