### 2025-08-20 #8828 @rahulkhorana
Hello @JacksonBurns and @ritesh001,

I would like to mention that the repository provides a dedicated issue template for reviews ([Issues](https://github.com/rahulkhorana/PolyatomicComplexes/issues)), which you are most welcome to use if it is convenient during your review. Feedback here in the thread is of course equally welcome, so please feel free to use whichever approach is easiest for you.

Thank you very much for your time and support with this review.
