### 2025-09-07 8a92f11
render grouped citations in an actual group
