### 2025-09-07 rahulkhorana/PolyatomicComplexes#162 @JacksonBurns
[Review] @JacksonBurns Review Comments
### Strengths

Very thorough literature review and context, helpful in conducting this review.

The colab demo is also very thorough, and helpful in understanding the functionality of `PolyatomicComplexes`. It ran all the way to the end without issue, and even included a nice meme.

# Code

In multiple places you modify the system path:
```python
BASE_PATH = Path(__file__)
project_root = BASE_PATH.parent.parent.parent.parent.resolve()
src_dir = project_root
if str(src_dir) not in sys.path:
    sys.path.insert(0, str(src_dir))
```

As a user, I would be rather surprised to find that a package which I had installed was modifying my `PATH` silently. At a minimum, some debugging information should be printed to at least _say_ that this is happening. I would personally prefer that this isn't done at all - why _is_ it currently done?

For the required Python version, you have `3.11.11` for installing with `pip`,  `python >= 3.11.11` when installing with the repo, and the `setup.py` does not actually limit the Python version at all. Please unify the suggested/required version, and further consider expanding it. You can almost definitely support all patch releases of Python 3.11 (i.e., `3.11.*`), and it would help in adoption if you could also support 3.12, at least.

# Documentation

 - The Table of Contents on the documentation landing page is not helpful, as it just shows the entire content of the bulleted list: https://rahulkhorana.github.io/PolyatomicComplexes/
 - The Usage Example for the `AbstractComplex` does not actually show anything about the `AbstractComplex` class itself (https://rahulkhorana.github.io/PolyatomicComplexes/abstract_complex/). It would be more useful to show how to implement one's own class by inheriting from this class.
 - Why is `from polyatomic_complexes.src.complexes import *` required in this demo (https://rahulkhorana.github.io/PolyatomicComplexes/datasets/)? `*` imports are generally not advisable.

# Paper

 - Broadly speaking, lines 18 - 45 restate the conclusions of your 2024 paper. I would consider just citing that paper (i.e. retaining only lines 18 and 45) and referencing your other work. If not:
    - Line 20: `Invariances: Representations should be invariant under changes in atom indexing and molecular rotation, reflection, and translation (Langer et al., 2022).`
    - Line 23: Drop `Essentially`
    - Line 28: This is pulled directly from your 2024 paper (`Moreover, discontinuities work against
regularity assumptions of many machine learning models`) - you may wish to add a citation there.
    - Line 29: Missing a period (`.`)
 - Lines 90 to 99: Add ticks (i.e. \`) around any class names so that they are rendered in `CodeFormat`

### Technical checks

You should verify that your continuous integration still passes (it appears the last run was 7 months ago) and you should add a scheduled runner with something like this: https://github.com/JacksonBurns/astartes/blob/main/.github/workflows/ci.yml#L2-L4 which will catch software rot

The use of `from .etc import *` all over the repo's `__init__.py` files makes it possible to `import` any class directly from the parent module:

```python
# demo suggests:
# from polyatomic_complexes.src.complexes import PolyatomicGeometrySMILE
# from polyatomic_complexes.src.complexes.abstract_complex import AbstractComplex
# but this is possible:
from polyatomic_complexes import AbstractComplex
from polyatomic_complexes import PolyatomicGeometrySMILE
```

I imagine this is not the desired effect, but maybe for some classes with deeply nested import statements I would consider using it. I mention this not with a particular suggestion in mind, but as a 'heads up'.

### Additional comments

I've opened a pull request to fix a formatting issue with the paper: https://github.com/rahulkhorana/PolyatomicComplexes/pull/161

I suggest resolving your outstanding dependabot PRs

You license year is out of date (2024).

You should add a `README.md` to `polyatomic_complexes/dataset` and add some provenance information for the various CSVs included therein (where you got them (URLs), citations, etc.).
