### 2025-09-07 #8828 @JacksonBurns
@rahulkhorana I've made a review issue here: https://github.com/rahulkhorana/PolyatomicComplexes/issues/162 using your suggested template (was helpful, thanks!)

An overall note, which you may take or leave at your discretion: consider running an autoformatter such as `black` on your repository (see also `isort`, `autoflake`, and others). This will make things 'prettier' for others to read.
