### 2025-09-08 rahulkhorana/PolyatomicComplexes#163 @rahulkhorana
JOSS Review [1] - code & path
In multiple places you modify the system path:

```python
BASE_PATH = Path(__file__)
project_root = BASE_PATH.parent.parent.parent.parent.resolve()
src_dir = project_root
if str(src_dir) not in sys.path:
    sys.path.insert(0, str(src_dir))
```

As a user, I would be rather surprised to find that a package which I had installed was modifying my PATH silently. At a minimum, some debugging information should be printed to at least say that this is happening. I would personally prefer that this isn't done at all - why is it currently done?

For the required Python version, you have 3.11.11 for installing with pip, python >= 3.11.11 when installing with the repo, and the setup.py does not actually limit the Python version at all. Please unify the suggested/required version, and further consider expanding it. You can almost definitely support all patch releases of Python 3.11 (i.e., 3.11.*), and it would help in adoption if you could also support 3.12, at least.
