### 2025-09-14 rahulkhorana/PolyatomicComplexes#167 @ritesh001
[Review] @ritesh001 Review Comments
### Strengths

The overall codebase is well-structured. Here are my assessment of the strenghs of the code:

  - Comprehensive tests with fuzzing
      - Many tests validate structure, uniqueness across molecules, and invariants
  - Multiple representations supported
      - graph kernels; fingerprints/SELFIES/SMILES encodings; graph GP pipelines integrated.
  - CI and packaging basics in place
      - GitHub Actions for build/CI with LFS pulls for datasets; `setup.py` packages datasets for out‑of‑the‑box examples; README with install paths and Colab link.

### Suggestions for improvement

Minor improvements:
1. Persistence builder skips last chunk and can set invalid chunk_size
  - In get_persistence_features, chunk building uses `range(num_chunks - 1)` which skips the last chunk. Also, chunk_size can become 0, causing a divide-by-zero or zero-chunk run. It can lead to silent data loss during persistence computation and potential runtime errors.
  - Proposed fix:
      - Can ensure `chunk_size = max(1, computed_value)` and keep a stable upper bound.
      - Can iterate over all chunks: `range(num_chunks)`.
      - Prefer sparse matrices to avoid `.toarray()` when possible.
  - Code pointers:
      - `polyatomic_complexes/src/complexes/space_utils.py`, `geometricPolyatomicComplex.get_persistence_features` and `geometricAtomicComplex.get_persistence_features`.

2. Block sum with wrong dimensional assumptions
   - `PolyAtomComplex.dir_sum` assumes 2D shapes; `self.F/self.DE` start as 1D empty arrays. This can error when `using_force`/`using_radial` are enabled.
   - Code-pointers:
       - `src/complexes/polyatomic_complex.py`.

3. Wrong return type in utility
      - `GeneralComplexUtils.get_nsphere_and_limits` returns `Tuple[nsphere, boundary]` (typing alias), not an actual tuple. Should return `(nsphere, boundary)`.
      - Code pointers:
         - `src/complexes/general_utils.py`.

### Technical checks

None

### Additional comments

_No response_
