### 2025-09-15 f519969
addresses issue #163
