+++ polyatomic_complexes/src/complexes/space_utils.py
@@ -16,6 +16,7 @@ project_root = BASE_PATH.parent.parent.parent.parent.resolve()
 src_dir = project_root
 if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
+print("Loading [polyatomic-complexes]")
 
 from polyatomic_complexes.src.complexes.building_blocks import Neutron, Proton, Electron
 
