+++ polyatomic_complexes/src/complexes/atomic_complex.py
@@ -10,6 +10,8 @@ src_dir = project_root
 if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
+print("Loading [polyatomic-complexes]")
+
 from polyatomic_complexes.src.complexes.building_blocks import Electron, Proton, Neutron
 from polyatomic_complexes.src.complexes.general_utils import GeneralComplexUtils
 
