+++ polyatomic_complexes/src/complexes/general_utils.py
@@ -13,6 +13,8 @@ src_dir = project_root
 if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
+print("Loading [polyatomic-complexes]")
+
 from polyatomic_complexes.src.complexes.core_utils import GluingMap, NSphere
 
 
