+++ polyatomic_complexes/src/complexes/polyatomic_complex.py
@@ -15,6 +15,8 @@ src_dir = project_root
 if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
+print("Loading [polyatomic-complexes]")
+
 from polyatomic_complexes.src.complexes.core_utils import GluingMap, ElectronField
 from polyatomic_complexes.src.complexes.building_blocks import Electron
 
