+++ polyatomic_complexes/src/complexes/build_atom_lookup.py
@@ -12,6 +12,8 @@ src_dir = project_root
 if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
+print("Loading [polyatomic-complexes]")
+
 from polyatomic_complexes.src.complexes.atomic_complex import AtomComplex
 
 
