### 2025-09-15 22bd5fe
resolves issue #165
