### 2025-09-15 470cdd4
should mostly handle #166
