+++ polyatomic_complexes/experiments/metrics.py
@@ -3,7 +3,7 @@ import numpy as np
 
 
 class CRPS:
-    def __init__(self, x_test: torch.Tensor, y_test: torch.Tensor) -> torch.Tensor:
+    def __init__(self, x_test: torch.Tensor, y_test: torch.Tensor):
         self.pred = torch.from_numpy(x_test)
         self.truth = torch.from_numpy(y_test)
 
@@ -18,19 +18,21 @@ class CRPS:
         :rtype: torch.Tensor
         :used per the Apache-2.0 license
         """
-        opts = dict(device=self.pred.device, dtype=self.pred.dtype)
         num_samples = self.pred.size(0)
         if num_samples == 1:
             return (self.pred[0] - self.truth).abs()
 
         self.pred = self.pred.sort(dim=0).values
         diff = self.pred[1:] - self.pred[:-1]
-        weight = torch.arange(1, num_samples, **opts) * torch.arange(
-            num_samples - 1, 0, -1, **opts
+        weight = torch.arange(
+            1, num_samples, dtype=self.pred.dtype, device=self.pred.device
+        ) * torch.arange(
+            num_samples - 1, 0, -1, dtype=self.pred.dtype, device=self.pred.device
         )
         assert isinstance(weight, torch.Tensor)
         weight = weight.reshape(weight.shape + (1,) * (diff.dim() - 1))
 
-        return (self.pred - self.truth).abs().mean(0) - (diff * weight).sum(
+        crps = (self.pred - self.truth).abs().mean(0) - (diff * weight).sum(
             0
         ) / num_samples**2
+        return crps
