### 2025-09-15 e3bb9a7
resolves #167
