### 2025-09-15 rahulkhorana/PolyatomicComplexes#167 @rahulkhorana
Hello!

I have resolved these issues with commit `edbb9a7`.

1. now uses `bound = max(1, chunk_size)`
2. now explicitly accounts in `dir_sum` for case where a and b are 1D with if cond
3. fixed return type
4. The other reviewer didn't seem to have an issue with the notebook.... are you using a high ram instance + 24 hr session? some of the computations can take a while.
