+++ requirements/requirements_graph.txt
@@ -175,7 +175,7 @@ selfies==2.1.2
 Send2Trash==1.8.3
 six==1.17.0
 sniffio==1.3.1
-soupsieve==2.6
+soupsieve==2.8
 spglib==2.5.0
 SQLAlchemy==2.0.37
 stack-data==0.6.3
