### 2025-09-15 4fd2bc2
Bump jupyter-core from 5.7.2 to 5.8.1
Bumps [jupyter-core](https://github.com/jupyter/jupyter_core) from 5.7.2 to 5.8.1.
- [Release notes](https://github.com/jupyter/jupyter_core/releases)
- [Changelog](https://github.com/jupyter/jupyter_core/blob/main/CHANGELOG.md)
- [Commits](https://github.com/jupyter/jupyter_core/compare/v5.7.2...v5.8.1)

---
updated-dependencies:
- dependency-name: jupyter-core
  dependency-version: 5.8.1
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
