+++ requirements/requirements_graph.txt
@@ -81,7 +81,7 @@ jsonschema-specifications==2024.10.1
 jupyter-events==0.11.0
 jupyter-lsp==2.2.5
 jupyter_client==8.6.3
-jupyter_core==5.7.2
+jupyter_core==5.8.1
 jupyter_server==2.15.0
 jupyter_server_terminals==0.5.3
 jupyterlab==4.3.4
