### 2025-09-15 9920d10
Bump parso from 0.8.4 to 0.8.5
Bumps [parso](https://github.com/davidhalter/parso) from 0.8.4 to 0.8.5.
- [Changelog](https://github.com/davidhalter/parso/blob/master/CHANGELOG.rst)
- [Commits](https://github.com/davidhalter/parso/compare/v0.8.4...v0.8.5)

---
updated-dependencies:
- dependency-name: parso
  dependency-version: 0.8.5
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
