+++ requirements/requirements.txt
@@ -100,7 +100,7 @@ multipledispatch==1.0.0
 mypy==1.14.1
 mypy-extensions==1.0.0
 nbclient==0.10.2
-nbconvert==7.16.5
+nbconvert==7.16.6
 nbformat==5.10.4
 nest-asyncio==1.6.0
 networkx==3.4.2
