### 2025-09-15 1a2a01d
Bump pydantic-core from 2.27.2 to 2.39.0
Bumps [pydantic-core](https://github.com/pydantic/pydantic-core) from 2.27.2 to 2.39.0.
- [Release notes](https://github.com/pydantic/pydantic-core/releases)
- [Commits](https://github.com/pydantic/pydantic-core/compare/v2.27.2...v2.39.0)

---
updated-dependencies:
- dependency-name: pydantic-core
  dependency-version: 2.39.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
