### 2025-09-15 98febe5
Bump pyopengl-accelerate from 3.1.9 to 3.1.10
Bumps [pyopengl-accelerate](https://github.com/mcfletch/pyopengl) from 3.1.9 to 3.1.10.
- [Commits](https://github.com/mcfletch/pyopengl/commits/3.1.10)

---
updated-dependencies:
- dependency-name: pyopengl-accelerate
  dependency-version: 3.1.10
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
