+++ requirements/requirements_mat.txt
@@ -148,7 +148,7 @@ pyglview==1.3.4
 Pygments==2.19.1
 pymatgen==2025.1.24
 PyOpenGL==3.1.9
-PyOpenGL-accelerate==3.1.9
+PyOpenGL-accelerate==3.1.10
 pyparsing==3.2.1
 pyro-api==0.1.2
 pyro-ppl==1.9.1
