+++ requirements/requirements.txt
@@ -167,7 +167,7 @@ six==1.17.0
 sniffio==1.3.1
 soupsieve==2.6
 spglib==2.5.0
-SQLAlchemy==2.0.37
+SQLAlchemy==2.0.43
 stack-data==0.6.3
 suds-community==1.2.0
 sympy==1.13.1
