### 2025-09-15 6e870cd
Bump pyscf from 2.8.0 to 2.10.0
Bumps [pyscf](https://github.com/pyscf/pyscf) from 2.8.0 to 2.10.0.
- [Release notes](https://github.com/pyscf/pyscf/releases)
- [Changelog](https://github.com/pyscf/pyscf/blob/master/CHANGELOG)
- [Commits](https://github.com/pyscf/pyscf/compare/v2.8.0...v2.10.0)

---
updated-dependencies:
- dependency-name: pyscf
  dependency-version: 2.10.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
