+++ requirements/requirements_graph.txt
@@ -19,7 +19,7 @@ botorch==0.12.0
 cattrs==24.1.2
 certifi==2024.12.14
 cffi==1.17.1
-charset-normalizer==3.4.1
+charset-normalizer==3.4.3
 click==8.1.8
 colorama==0.4.6
 colorlog==6.9.0
