### 2025-09-15 59991e0
Bump argon2-cffi-bindings from 21.2.0 to 25.1.0
Bumps [argon2-cffi-bindings](https://github.com/hynek/argon2-cffi-bindings) from 21.2.0 to 25.1.0.
- [Release notes](https://github.com/hynek/argon2-cffi-bindings/releases)
- [Changelog](https://github.com/hynek/argon2-cffi-bindings/blob/main/CHANGELOG.md)
- [Commits](https://github.com/hynek/argon2-cffi-bindings/compare/21.2.0...25.1.0)

---
updated-dependencies:
- dependency-name: argon2-cffi-bindings
  dependency-version: 25.1.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] 
[…]
