+++ requirements/requirements.txt
@@ -3,7 +3,7 @@ anyio==4.8.0
 appdirs==1.4.4
 appnope==0.1.4
 argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
+argon2-cffi-bindings==25.1.0
 arrow==1.3.0
 ase==3.24.0
 asttokens==3.0.0
