### 2025-09-15 fa623bd
Update ci.yml with longer timeout
