+++ .github/workflows/ci.yml
@@ -10,6 +10,7 @@ on:
 jobs:
   test:
     runs-on: ${{ matrix.os }}
+    timeout-minutes: 720
     strategy:
       matrix:
         os: [ubuntu-latest]
