### 2025-09-15 a0154b8
Bump the pip group across 1 directory with 6 updates
Bumps the pip group with 6 updates in the /requirements directory:

| Package | From | To |
| --- | --- | --- |
| [deepdiff](https://github.com/seperman/deepdiff) | `8.1.1` | `8.6.1` |
| [h11](https://github.com/python-hyper/h11) | `0.14.0` | `0.16.0` |
| [requests](https://github.com/psf/requests) | `2.32.3` | `2.32.4` |
| [torch](https://github.com/pytorch/pytorch) | `2.5.1` | `2.8.0` |
| [tornado](https://github.com/tornadoweb/tornado) | `6.4.2` | `6.5` |
| [urllib3](https://github.com/urllib3/urllib3) | `2.3.0` | `2.5.0` |



Updates `de
[…]
