+++ requirements/requirements_mat.txt
@@ -31,7 +31,7 @@ cycler==0.12.1
 Cython==3.0.11
 debugpy==1.8.11
 decorator==5.1.1
-deepdiff==8.1.1
+deepdiff==8.6.1
 defusedxml==0.7.1
 Deprecated==1.2.15
 dftd3==1.2.1
@@ -58,7 +58,7 @@ graphein==1.7.7
 greenlet==3.1.1
 grequests==0.7.0
 gudhi==3.10.1
-h11==0.14.0
+h11==0.16.0
 h5py==3.12.1
 httpcore==1.0.7
 httpx==0.28.1
@@ -162,7 +162,7 @@ PyYAML==6.0.2
 pyzmq==26.2.0
 rdkit==2024.9.4
 referencing==0.35.1
-requests==2.32.3
+requests==2.32.4
 requests-cache==1.2.1
 rfc3339-validator==0.1.4
 rfc3986-validator==0.1.1
@@ -190,8 +190,8 @@ terminado==0.18.1
 threadpoolctl==3.5.0
 tinycss2==1.4.0
 TopoNetX==0.2.0
-torch==2.5.1
-tornado==6.4.2
+torch==2.8.0
+tornado==6.5
 tqdm==4.67.1
 traitlets==5.14.3
 trimesh==4.5.3
@@ -202,7 +202,7 @@ tzdata==2024.2
 uncertainties==3.2.2
 uri-template==1.3.0
 url-normalize==1.4.3
-urllib3==2.3.0
+urllib3==2.5.0
 wcwidth==0.2.13
 webcolors==24.11.1
 webencodings==0.5.1
