+++ .github/workflows/ci.yml
@@ -18,10 +18,9 @@ jobs:
         os: [ubuntu-latest]
         python-version: ['3.11.11']
         shard_total: [8]
-        shard_index: [1,2,3,4,5,6,7,8]
+        shard_index: [0,1,2,3,4,5,6,7]
 
     env:
-      PYTEST_DISABLE_PLUGIN_AUTOLOAD: "1"
       PYTHONOPTIMIZE: "1"
       XLA_PYTHON_CLIENT_PREALLOCATE: "false"
       PYTEST_ADDOPTS: "--log-level=WARNING"
@@ -55,31 +54,24 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements/requirements.txt
-          pip install pytest-split
 
-      - name: Restore pytest-split durations cache
-        uses: actions/cache@v4
-        with:
-          path: polyatomic_complexes/.pytest-split-durations
-          key: durations-${{ github.ref_name }}
-          restore-keys: |
-            durations-
-
-      - name: Run tests (timed shard)
+      - name: Find and shard test files
         working-directory: ./polyatomic_complexes
         run: |
-          python -B -m pytest -p pytest_split.plugin \
-            --splits ${{ matrix.shard_total }} \
-            --group  ${{ matrix.shard_index }} \
-            --store-durations \
-            -v --maxfail=1
+          find . -type f -name "test_*.py" | sort > all_tests.txt
+          TOTAL=${{ matrix.shard_total }}
+          INDEX=${{ matrix.shard_index }}
+          awk "NR % $TOTAL == $INDEX" all_tests.txt > this_shard.txt
+          echo "Shard $INDEX/$TOTAL will run $(wc -l < this_shard.txt) files"
 
-      - name: Save updated durations cache
-        if: always()
-        uses: actions/cache@v4
-        with:
-          path: polyatomic_complexes/.pytest-split-durations
-          key: durations-${{ github.run_id }}
+      - name: Run shard
+        working-directory: ./polyatomic_complexes
+        run: |
+          split -l 20 this_shard.txt shard_
+          for f in shard_*; do
+            echo "=== Running batch $f ==="
+            python -B -m pytest -v --maxfail=1 $(cat "$f") || exit 1
+          done
 
       - name: Upload coverage reports to Codecov
         if: always()
