+++ .github/workflows/ci.yml
@@ -7,99 +7,73 @@ on:
 
 jobs:
   test:
-    name: test (shard ${{ matrix.shard_index }}/${{ matrix.shard_total }})
+    name: test (shard ${{ matrix.shard_index + 1 }}/${{ strategy.matrix.shard_total }}) # <--- Made this 1-indexed for readability
     runs-on: ubuntu-latest
     timeout-minutes: 90
     strategy:
       fail-fast: false
       matrix:
-        python-version: ['3.11.11']
+        python-version: ['3.11'] # <--- Simplified version, setup-python handles specifics
         shard_total: [8]
-        shard_index: [0,1,2,3,4,5,6,7]
+        # <--- Use range() for a cleaner way to define indices
+        shard_index: ${{ toJSON(range(8)) }} 
 
     env:
       PYTHONOPTIMIZE: "0"
       PYTEST_ADDOPTS: "--log-level=WARNING"
 
     steps:
-      - name: Checkout code
+      - name: Checkout code with LFS
         uses: actions/checkout@v4
-
-      - name: Install Git LFS
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y git-lfs
-          git lfs install
+        with:
+          lfs: true # <--- CHANGE: Built-in LFS handling is much cleaner
 
       - name: Setup Python
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
 
-      - name: Verify Python version
-        run: python --version
-
-      - name: Pull LFS dataset
-        run: |
-          git lfs pull --include=polyatomic_complexes/dataset/**/*.csv
-          git lfs pull --include=polyatomic_complexes/dataset/**/*.json
-          git lfs pull --include=polyatomic_complexes/dataset/**/atom_lookup.pkl
-
       - name: Install dependencies
         run: |
           pip install -r requirements/requirements.txt
+          pip install pytest-xdist # <--- OPTIONAL but recommended for even better splitting
 
-      - name: Collect static test nodeids (no imports, no OOM)
+      - name: Collect and shard tests
+        id: collect_tests # <--- Give the step an ID to reference its output
         working-directory: ./polyatomic_complexes
         run: |
-          echo "Generating static test list..."
-          find . -type f -name "test_*.py" | sort > all_test_files.txt
+          # <--- CHANGE: Use pytest's robust collection, not fragile shell parsing
+          pytest --collect-only -q > all_nodeids.txt
+          
+          echo "Total tests found: $(wc -l < all_nodeids.txt)"
 
-          > all_nodeids.txt
-          while read file; do
-            while read -r line; do
-              if [[ $line =~ ^def\ (test_[a-zA-Z0-9_]+) ]]; then
-                test_name="${BASH_REMATCH[1]}"
-                echo "${file#./}::${test_name}" >> all_nodeids.txt
-              fi
-            done < "$file"
-          done < all_test_files.txt
+          # Shard the collected test node IDs
+          awk "NR % ${{ matrix.shard_total }} == ${{ matrix.shard_index }}" all_nodeids.txt > shard.txt
 
-          echo "Total tests found:"
-          wc -l all_nodeids.txt
-
-      - name: Shard test nodeids
-        working-directory: ./polyatomic_complexes
-        run: |
-          TOTAL=${{ matrix.shard_total }}
-          INDEX=${{ matrix.shard_index }}
-
-          awk "NR % $TOTAL == $INDEX" all_nodeids.txt > shard.txt
-
-          echo "Shard $INDEX/$TOTAL will run $(wc -l < shard.txt) test(s)."
-          cat shard.txt || echo "No tests assigned."
-
-      - name: Run this shard
-        working-directory: ./polyatomic_complexes
-        run: |
[…]
