+++ .github/workflows/ci.yml
@@ -2,67 +2,53 @@ name: ci
 
 on:
   push:
-    branches: [main, productionizing]
+    branches:
+      - main
+      - productionizing
   workflow_dispatch:
 
 jobs:
   test:
-    name: test (shard ${{ strategy.matrix.shard }}/${{ strategy.matrix.shard_total }})
-    runs-on: ubuntu-latest
-    timeout-minutes: 90
+    runs-on: ${{ matrix.os }}
     strategy:
-      fail-fast: false
       matrix:
-        python-version: ['3.11']
-        shard_total: [8]
-        # Create a sequence from 1 to 8 for the shard index
-        shard: [1, 2, 3, 4, 5, 6, 7, 8]
+        os: [ubuntu-latest]
+        python-version: ['3.11.11']
 
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-        with:
-          # A full clone is often required for accurate test discovery
-          fetch-depth: 0
-          lfs: true
+      - name: Checkout repository content
+        uses: actions/checkout@v3
+
+      - name: Install Git LFS
+        run: |
+          sudo apt-get install git-lfs
+          git lfs install
 
       - name: Setup Python
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+          cache: 'pip'
 
-      - name: Install dependencies
+      - name: Verify Python version
+        run: python --version
+
+      - name: Checkout LFS
         run: |
+          git lfs pull --include=polyatomic_complexes/dataset/**/*.csv
+          git lfs pull --include=polyatomic_complexes/dataset/**/*.json
+          git lfs pull --include=polyatomic_complexes/dataset/**/atom_lookup.pkl
+      
+      - name: Install Python packages
+        run: |
+          python -m pip install --upgrade pip
           pip install -r requirements/requirements.txt
-          # Install the plugin that handles test splitting
-          pip install pytest-split
 
-      - name: Run test shard
-        run: |
-          # pytest-split handles the collection and splitting automatically.
-          # Each shard runs its dedicated fraction of the tests.
-          # Add --cov flags if you are measuring coverage.
-          pytest \
-            --num-groups ${{ matrix.shard_total }} \
-            --group ${{ matrix.shard }}
+      - name: Run tests
+        working-directory: ./polyatomic_complexes
+        run: python -B -m pytest -v --log-level=DEBUG --maxfail=1
 
-      - name: Upload coverage to Codecov
-        if: always()
-        uses: codecov/codecov-action@v4
+      - name: Upload coverage reports to Codecov
+        uses: codecov/codecov-action@v5
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
-          flags: shard-${{ matrix.shard }}
-          fail_ci_if_error: false
-          verbose: true
-
-  codecov-finalize:
-    name: codecov finalize
-    needs: test
-    runs-on: ubuntu-latest
-    steps:
-      - name: Finalize Codecov
-        uses: codecov/codecov-action@v4
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }}
-          fail_ci_if_error: false
-          verbose: true
