### 2025-09-17 bf35f61
Update test_large_geom.py
