### 2025-09-17 0518e83
Update test_polyatoms.py
