+++ .github/workflows/ci.yml
@@ -8,47 +8,66 @@ on:
   workflow_dispatch:
 
 jobs:
-  test:
-    runs-on: ${{ matrix.os }}
+  test-shards:
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
     strategy:
       matrix:
-        os: [ubuntu-latest]
         python-version: ['3.11.11']
+        shard_total: [4]
+        shard_index: [0,1,2,3]
+
+    env:
+      PYTHONOPTIMIZE: "0"
+      PYTEST_DISABLE_PLUGIN_AUTOLOAD: "1"
+      PYTEST_ADDOPTS: "--log-level=WARNING"
 
     steps:
-      - name: Checkout repository content
-        uses: actions/checkout@v3
-
-      - name: Install Git LFS
-        run: |
-          sudo apt-get install git-lfs
-          git lfs install
+      - name: Checkout code
+        uses: actions/checkout@v4
 
       - name: Setup Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-          cache: 'pip'
 
-      - name: Verify Python version
-        run: python --version
-
-      - name: Checkout LFS
+      - name: Install dependencies
         run: |
-          git lfs pull --include=polyatomic_complexes/dataset/**/*.csv
-          git lfs pull --include=polyatomic_complexes/dataset/**/*.json
-          git lfs pull --include=polyatomic_complexes/dataset/**/atom_lookup.pkl
-      
-      - name: Install Python packages
-        run: |
-          python -m pip install --upgrade pip
           pip install -r requirements/requirements.txt
+
+      - name: Static nodeid generation (no dataset loads)
+        working-directory: ./polyatomic_complexes
+        run: |
+          find . -type f -name "test_*.py" | sort > all_test_files.txt
+
+          > all_nodeids.txt
+          while read file; do
+            # Only grep for definitions, not import & execute
+            grep -E '^def test_' "$file" | sed "s/^def //; s/(.*//" | while read testfn; do
+              echo "${file#./}::${testfn}" >> all_nodeids.txt
+            done
+          done < all_test_files.txt
+
+          echo "Found $(wc -l < all_nodeids.txt) test functions"
+
+      - name: Pick shard
+        working-directory: ./polyatomic_complexes
+        run: |
+          TOTAL=${{ matrix.shard_total }}
+          INDEX=${{ matrix.shard_index }}
+          awk "NR % $TOTAL == $INDEX" all_nodeids.txt > shard.txt
+          echo "Shard $INDEX/$TOTAL → $(wc -l < shard.txt) tests"
+          cat shard.txt
 
       - name: Run tests
         working-directory: ./polyatomic_complexes
-        run: python -B -m pytest -v --log-level=DEBUG --maxfail=1
-
-      - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v5
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }}
+        run: |
+          if [ -s shard.txt ]; then
+            split -l 50 shard.txt batch_
+            for f in batch_*; do
+              echo "Running batch $f"
+              python -B -m pytest -v --maxfail=1 $(cat "$f") || exit 1
+            done
+          else
+            echo "No tests for this shard"
+          fi
