+++ .github/workflows/ci.yml
@@ -25,10 +25,18 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
+          cache: pip
+
+      - name: Cache pip dependencies
+        uses: actions/cache@v4
+        with:
+          path: ~/.cache/pip
+          key: pip-${{ hashFiles('requirements/requirements.txt') }}
+          restore-keys: |
+            pip-
 
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
           pip install -r requirements/requirements.txt
 
       - name: Collect test nodeids (static scan)
