+++ polyatomic_complexes/tests/test_large_geom.py
@@ -22,14 +22,14 @@ from polyatomic_complexes.src.complexes.abstract_complex import AbstractComplex
 from polyatomic_complexes.src.complexes.quantum_theor_complex import QuantumWavesComplex
 
 
-from .test_geom import (
+from polyatomic_complexes.tests.test_geom import (
     check_abstract_complex_test,
     check_force_complex_test,
     check_quantum_complex_test,
     check_quantum_waves_complex_test,
 )
 
-from .test_geom_unique import (
+from polyatomic_complexes.tests.test_geom_unique import (
     check_adjacency_lists_unique,
     check_betti_numbers_unique,
     check_incidence_unique,
@@ -53,7 +53,7 @@ smiles_freesolv = pd.read_csv(datapath_freesolv)["smiles"].tolist()
 smiles_lipo = pd.read_csv(datapath_lipo)["smiles"].tolist()
 smiles_photo = pd.read_csv(datapath_photo)["SMILES"].tolist()
 ALL_SMILES = smiles_esol + smiles_freesolv + smiles_lipo + smiles_photo
-ALL_SMILES = np.random.choice(a=ALL_SMILES, size=5, replace=False).tolist()
+ALL_SMILES = np.random.choice(a=ALL_SMILES, size=2, replace=False).tolist()
 
 large_cases = []
 for smile in ALL_SMILES[:50]:
@@ -94,8 +94,8 @@ def test_quantum_polyatomic_geometry(smile, mode):
 
 #### LARGE UNIQUE ####
 large_cases_pair = []
-for smile1 in ALL_SMILES[:5]:
-    for smile2 in ALL_SMILES[:5]:
+for smile1 in ALL_SMILES:
+    for smile2 in ALL_SMILES:
         if smile1 != smile2:
             for mode in modes:
                 large_cases_pair.append((smile1, smile2, mode))
