### 2025-09-21 #8828 @rahulkhorana
Hi @RMeli,

Thanks again for your help with this submission.

Since both reviews are complete and I’ve addressed the suggested changes ([#162](https://github.com/rahulkhorana/PolyatomicComplexes/issues/162), [#167](https://github.com/rahulkhorana/PolyatomicComplexes/issues/167)), I just wanted to check if there’s anything further required from my side.

Looping in @crvernon as well.

Thanks so much for your time and help!

Best,
Rahul
