+++ polyatomic_complexes/tests/test_kchains.py
@@ -142,7 +142,8 @@ smiles_freesolv = pd.read_csv(datapath_freesolv)["smiles"].tolist()
 smiles_lipo = pd.read_csv(datapath_lipo)["smiles"].tolist()
 smiles_photo = pd.read_csv(datapath_photo)["SMILES"].tolist()
 ALL_SMILES = smiles_esol + smiles_freesolv + smiles_lipo + smiles_photo
-ALL_SMILES = np.random.choice(a=ALL_SMILES, size=2, replace=False).tolist()
+ALL_SMILES = np.random.choice(a=ALL_SMILES, size=100, replace=False).tolist()
+ALL_SMILES = [s for s in ALL_SMILES if len(s) < 7]
 
 
 cases_pair_large = []
