### 2025-09-21 51a7e9b
adding cases
