+++ polyatomic_complexes/tests/test_large_geom.py
@@ -54,10 +54,10 @@ smiles_lipo = pd.read_csv(datapath_lipo)["smiles"].tolist()
 smiles_photo = pd.read_csv(datapath_photo)["SMILES"].tolist()
 ALL_SMILES = smiles_esol + smiles_freesolv + smiles_lipo + smiles_photo
 ALL_SMILES = np.random.choice(a=ALL_SMILES, size=1000, replace=False).tolist()
-ALL_SMILES = [s for s in ALL_SMILES if len(s) < 10]
+ALL_SMILES = [s for s in ALL_SMILES if len(s) < 10][:10]
 
 large_cases = []
-for smile in ALL_SMILES[:50]:
+for smile in ALL_SMILES:
     for mode in modes:
         large_cases.append((smile, mode))
 
