### 2025-09-21 16e8e8d
further mem
