### 2025-09-21 5f9c730
check smiles are valid
