+++ polyatomic_complexes/tests/test_large_geom.py
@@ -3,6 +3,7 @@ import sys
 import pytest
 import numpy as np
 import pandas as pd
+from rdkit import Chem
 from pathlib import Path
 
 sys.path.append(
@@ -42,6 +43,12 @@ from polyatomic_complexes.tests.test_geom_unique import (
 
 modes = ["abstract", "force-field"]
 
+
+def is_valid_smiles(s):
+    mol = Chem.MolFromSmiles(s)
+    return mol is not None
+
+
 #### LARGE SANITY ####
 parent_path = BASE_PATH.parent.parent.__str__()
 datapath_esol = parent_path + "/dataset/esol/ESOL.csv"
@@ -54,7 +61,7 @@ smiles_lipo = pd.read_csv(datapath_lipo)["smiles"].tolist()
 smiles_photo = pd.read_csv(datapath_photo)["SMILES"].tolist()
 ALL_SMILES = smiles_esol + smiles_freesolv + smiles_lipo + smiles_photo
 ALL_SMILES = np.random.choice(a=ALL_SMILES, size=1000, replace=False).tolist()
-ALL_SMILES = [s for s in ALL_SMILES if len(s) < 10][:3]
+ALL_SMILES = [s for s in ALL_SMILES if len(s) < 10 and is_valid_smiles(s)][:3]
 
 large_cases = []
 for smile in ALL_SMILES:
