### 2025-09-22 rahulkhorana/PolyatomicComplexes#163 @rahulkhorana
@RMeli  @JacksonBurns,

Yeah good points! I can make the prints like formal debug messages if you want or info messages using logging if you wish.

To be honest you need to make the imports and path additions absolute so the importing works. This is part of PEP328 standard now: https://stackoverflow.com/questions/4209641/absolute-vs-explicit-relative-import-of-python-module.

Basically, its no longer best practice as per PEP328 hence the addition also so functionality.
