### 2025-09-23 45887c1
Bump pybtex from 0.24.0 to 0.25.1
Bumps [pybtex](https://bitbucket.org/pybtex-devs/pybtex) from 0.24.0 to 0.25.1.
- [Changelog](https://bitbucket.org/pybtex-devs/pybtex/src/master/CHANGES)
- [Commits](https://bitbucket.org/pybtex-devs/pybtex/commits)

---
updated-dependencies:
- dependency-name: pybtex
  dependency-version: 0.25.1
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
