+++ requirements/requirements.txt
@@ -127,7 +127,7 @@ psutil==6.1.1
 ptyprocess==0.7.0
 pure_eval==0.2.3
 pyarrow==18.1.0
-pybtex==0.24.0
+pybtex==0.25.1
 pycodestyle==2.12.1
 pycparser==2.22
 pydantic==2.10.5
