### 2025-09-23 aefb15b
Bump spglib from 2.5.0 to 2.6.0
Bumps [spglib](https://github.com/spglib/spglib) from 2.5.0 to 2.6.0.
- [Release notes](https://github.com/spglib/spglib/releases)
- [Changelog](https://github.com/spglib/spglib/blob/develop/ChangeLog.md)
- [Commits](https://github.com/spglib/spglib/compare/v2.5.0...v2.6.0)

---
updated-dependencies:
- dependency-name: spglib
  dependency-version: 2.6.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
