+++ requirements/requirements_graph.txt
@@ -176,7 +176,7 @@ Send2Trash==1.8.3
 six==1.17.0
 sniffio==1.3.1
 soupsieve==2.6
-spglib==2.5.0
+spglib==2.6.0
 SQLAlchemy==2.0.37
 stack-data==0.6.3
 suds-community==1.2.0
