### 2025-09-23 973e8e5
Bump pyzmq from 26.2.0 to 27.1.0
Bumps [pyzmq](https://github.com/zeromq/pyzmq) from 26.2.0 to 27.1.0.
- [Release notes](https://github.com/zeromq/pyzmq/releases)
- [Commits](https://github.com/zeromq/pyzmq/compare/v26.2.0...v27.1.0)

---
updated-dependencies:
- dependency-name: pyzmq
  dependency-version: 27.1.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
