+++ requirements/requirements_graph.txt
@@ -156,7 +156,7 @@ python-dateutil==2.9.0.post0
 python-json-logger==3.2.1
 pytz==2024.2
 PyYAML==6.0.2
-pyzmq==26.2.0
+pyzmq==27.1.0
 rdkit==2024.9.4
 referencing==0.35.1
 requests==2.32.3
