### 2025-09-23 abb4607
Bump fonttools from 4.55.3 to 4.60.0
Bumps [fonttools](https://github.com/fonttools/fonttools) from 4.55.3 to 4.60.0.
- [Release notes](https://github.com/fonttools/fonttools/releases)
- [Changelog](https://github.com/fonttools/fonttools/blob/main/NEWS.rst)
- [Commits](https://github.com/fonttools/fonttools/compare/4.55.3...4.60.0)

---
updated-dependencies:
- dependency-name: fonttools
  dependency-version: 4.60.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
