+++ requirements/requirements_mat.txt
@@ -42,7 +42,7 @@ executing==2.1.0
 fastjsonschema==2.21.1
 filelock==3.17.0
 flake8==7.1.1
-fonttools==4.55.3
+fonttools==4.60.0
 fqdn==1.5.1
 fsspec==2024.12.0
 future==1.0.0
