### 2025-09-23 269e348
Bump sympy from 1.13.1 to 1.14.0
Bumps [sympy](https://github.com/sympy/sympy) from 1.13.1 to 1.14.0.
- [Release notes](https://github.com/sympy/sympy/releases)
- [Commits](https://github.com/sympy/sympy/compare/sympy-1.13.1...sympy-1.14.0)

---
updated-dependencies:
- dependency-name: sympy
  dependency-version: 1.14.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
