+++ requirements/requirements_graph.txt
@@ -180,7 +180,7 @@ spglib==2.5.0
 SQLAlchemy==2.0.37
 stack-data==0.6.3
 suds-community==1.2.0
-sympy==1.13.1
+sympy==1.14.0
 tabulate==0.9.0
 tenacity==9.0.0
 terminado==0.18.1
