### 2025-09-23 4f6973c
Bump zope-event from 5.0 to 6.0
Bumps [zope-event](https://github.com/zopefoundation/zope.event) from 5.0 to 6.0.
- [Changelog](https://github.com/zopefoundation/zope.event/blob/master/CHANGES.rst)
- [Commits](https://github.com/zopefoundation/zope.event/compare/5.0...6.0)

---
updated-dependencies:
- dependency-name: zope-event
  dependency-version: '6.0'
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
