### 2025-09-23 1399437
Bump joblib from 1.4.2 to 1.5.2
Bumps [joblib](https://github.com/joblib/joblib) from 1.4.2 to 1.5.2.
- [Release notes](https://github.com/joblib/joblib/releases)
- [Changelog](https://github.com/joblib/joblib/blob/main/CHANGES.rst)
- [Commits](https://github.com/joblib/joblib/compare/1.4.2...1.5.2)

---
updated-dependencies:
- dependency-name: joblib
  dependency-version: 1.5.2
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
