+++ requirements/requirements.txt
@@ -72,7 +72,7 @@ jaxlib==0.4.38
 jaxtyping==0.2.19
 jedi==0.19.2
 Jinja2==3.1.5
-joblib==1.4.2
+joblib==1.5.2
 json5==0.10.0
 jsonpointer==3.0.0
 jsonschema==4.23.0
