+++ requirements/requirements.txt
@@ -83,7 +83,7 @@ line_profiler==4.2.0
 linear-operator==0.5.3
 loguru==0.7.3
 looseversion==1.1.2
-lxml==5.3.0
+lxml==6.0.2
 markdown-it-py==3.0.0
 MarkupSafe==3.0.2
 matplotlib==3.10.0
