### 2025-09-23 2543558
Bump biopython from 1.84 to 1.85
Bumps [biopython](https://github.com/biopython/biopython) from 1.84 to 1.85.
- [Changelog](https://github.com/biopython/biopython/blob/master/NEWS.rst)
- [Commits](https://github.com/biopython/biopython/commits)

---
updated-dependencies:
- dependency-name: biopython
  dependency-version: '1.85'
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
