### 2025-09-23 e58ac8e
Bump line-profiler from 4.2.0 to 5.0.0
Bumps [line-profiler](https://github.com/pyutils/line_profiler) from 4.2.0 to 5.0.0.
- [Release notes](https://github.com/pyutils/line_profiler/releases)
- [Changelog](https://github.com/pyutils/line_profiler/blob/main/CHANGELOG.rst)
- [Commits](https://github.com/pyutils/line_profiler/compare/v4.2.0...v5.0.0)

---
updated-dependencies:
- dependency-name: line-profiler
  dependency-version: 5.0.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
