+++ requirements/requirements.txt
@@ -79,7 +79,7 @@ jsonschema==4.23.0
 jsonschema-specifications==2024.10.1
 kiwisolver==1.4.8
 latexcodec==3.0.0
-line_profiler==4.2.0
+line_profiler==5.0.0
 linear-operator==0.5.3
 loguru==0.7.3
 looseversion==1.1.2
