+++ requirements/requirements_graph.txt
@@ -89,7 +89,7 @@ jupyterlab_pygments==0.3.0
 jupyterlab_server==2.27.3
 kiwisolver==1.4.8
 latexcodec==3.0.0
-line_profiler==4.2.0
+line_profiler==5.0.0
 linear-operator==0.5.3
 loguru==0.7.3
 looseversion==1.1.2
