### 2025-09-24 rahulkhorana/PolyatomicComplexes#163 @JacksonBurns
Seconding @RMeli, I'd rather see something like:

```python
BASE_PATH = Path(__file__)
project_root = BASE_PATH.parent.parent.parent.parent.resolve()
src_dir = project_root
if str(src_dir) not in sys.path:
    sys.path.insert(0, str(src_dir))
print(f"Prepended `{src_dir}` to PATH")
```

which makes it more explicit what's going on.

RE: absolute vs. relative imports

I don't really mind use of either of these, so please do whatever is best for you. My hold up is with the modification of the `PATH`. With the appropriate use of `__init__.py`'s and a `setup.py` file that correctly identifies all submodules, modifying the `PATH` shouldn't be necessary.

While we're talking about `import`s, I'll also mention that `from modulename import *` style "wildcard imports" should also be avoided, and there are some present in this codebase. From [PEP8](https://peps.python.org/pep-0008/#imports):

```
Wildcard imports (from <module> import *) should be avoided
```

I left this out of my initial review because, at the end of the day, it's a stylistic choice. But if we're going to possibly work on the imports across the codebase, it would be worth undoing this now IMO
