### 2025-09-24 #8828 @JacksonBurns
> Thank you [@ritesh001](https://github.com/ritesh001) and [@JacksonBurns](https://github.com/JacksonBurns) for completing your review so quickly! Could you please give me a brief summary of the review and if you are happy about how all the review processes?

The package is in excellent shape overall. The demos were especially good - they worked as promised and covered all of the key functionality for the package. The paper is well written and helpful, and the code seems largely correct and reasonably structured. My only substantial issues are with the [use of wildcard imports and PATH modification](https://github.com/rahulkhorana/PolyatomicComplexes/issues/163#issuecomment-3321508852), though I do not think these should be a barrier to acceptance if they are not resolved exactly to my liking.
