+++ requirements/requirements.txt
@@ -75,7 +75,7 @@ Jinja2==3.1.5
 joblib==1.4.2
 json5==0.10.0
 jsonpointer==3.0.0
-jsonschema==4.23.0
+jsonschema==4.25.1
 jsonschema-specifications==2024.10.1
 kiwisolver==1.4.8
 latexcodec==3.0.0
