+++ requirements/requirements.txt
@@ -120,7 +120,7 @@ pexpect==4.9.0
 pillow==11.1.0
 platformdirs==4.3.6
 plotly==5.24.1
-pluggy==1.5.0
+pluggy==1.6.0
 prometheus_client==0.21.1
 prompt_toolkit==3.0.48
 psutil==6.1.1
