+++ requirements/requirements_graph.txt
@@ -40,7 +40,7 @@ easydev==0.13.3
 easydict==1.13
 executing==2.1.0
 fastjsonschema==2.21.1
-filelock==3.17.0
+filelock==3.19.1
 flake8==7.1.1
 fonttools==4.55.3
 fqdn==1.5.1
