+++ requirements/requirements_mat.txt
@@ -196,7 +196,7 @@ tqdm==4.67.1
 traitlets==5.14.3
 trimesh==4.5.3
 typeguard==4.4.1
-types-python-dateutil==2.9.0.20241206
+types-python-dateutil==2.9.0.20250822
 typing_extensions==4.12.2
 tzdata==2024.2
 uncertainties==3.2.2
