### 2025-09-24 ba5f2d4
Bump rich-click from 1.8.5 to 1.9.1
Bumps [rich-click](https://github.com/ewels/rich-click) from 1.8.5 to 1.9.1.
- [Release notes](https://github.com/ewels/rich-click/releases)
- [Changelog](https://github.com/ewels/rich-click/blob/main/CHANGELOG.md)
- [Commits](https://github.com/ewels/rich-click/compare/v1.8.5...v1.9.1)

---
updated-dependencies:
- dependency-name: rich-click
  dependency-version: 1.9.1
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
