+++ requirements/requirements_graph.txt
@@ -164,7 +164,7 @@ requests-cache==1.2.1
 rfc3339-validator==0.1.4
 rfc3986-validator==0.1.1
 rich==13.9.4
-rich-click==1.8.5
+rich-click==1.9.1
 rpds-py==0.22.3
 ruamel.yaml==0.18.10
 ruamel.yaml.clib==0.2.12
