### 2025-09-24 3a652cf
Bump mistune from 3.1.0 to 3.1.4
Bumps [mistune](https://github.com/lepture/mistune) from 3.1.0 to 3.1.4.
- [Release notes](https://github.com/lepture/mistune/releases)
- [Changelog](https://github.com/lepture/mistune/blob/main/docs/changes.rst)
- [Commits](https://github.com/lepture/mistune/compare/v3.1.0...v3.1.4)

---
updated-dependencies:
- dependency-name: mistune
  dependency-version: 3.1.4
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
