+++ requirements/requirements.txt
@@ -90,7 +90,7 @@ matplotlib==3.10.0
 matplotlib-inline==0.1.7
 mccabe==0.7.0
 mdurl==0.1.2
-mistune==3.1.0
+mistune==3.1.4
 ml_dtypes==0.5.1
 mmtf-python==1.1.3
 monty==2025.1.9
