### 2025-09-24 b1c562b
Bump ase from 3.24.0 to 3.26.0
Bumps [ase](https://gitlab.com/ase/ase) from 3.24.0 to 3.26.0.
- [Changelog](https://gitlab.com/ase/ase/blob/master/CHANGELOG.rst)
- [Commits](https://gitlab.com/ase/ase/compare/3.24.0...3.26.0)

---
updated-dependencies:
- dependency-name: ase
  dependency-version: 3.26.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
