+++ requirements/requirements.txt
@@ -5,7 +5,7 @@ appnope==0.1.4
 argon2-cffi==23.1.0
 argon2-cffi-bindings==21.2.0
 arrow==1.3.0
-ase==3.24.0
+ase==3.26.0
 asttokens==3.0.0
 async-lru==2.0.4
 attrs==24.3.0
