+++ requirements/requirements.txt
@@ -10,7 +10,7 @@ asttokens==3.0.0
 async-lru==2.0.4
 attrs==24.3.0
 babel==2.16.0
-beautifulsoup4==4.12.3
+beautifulsoup4==4.13.5
 biopandas==0.5.1
 biopython==1.84
 bioservices==1.11.2
