+++ requirements/requirements_mat.txt
@@ -145,7 +145,7 @@ pydantic_core==2.27.2
 pyfiglet==1.0.2
 pyflakes==3.2.0
 pyglview==1.3.4
-Pygments==2.19.1
+Pygments==2.19.2
 pymatgen==2025.1.24
 PyOpenGL==3.1.9
 PyOpenGL-accelerate==3.1.9
