+++ requirements/requirements_graph.txt
@@ -144,7 +144,7 @@ pydantic==2.10.6
 pydantic_core==2.27.2
 pyfiglet==1.0.2
 pyflakes==3.2.0
-Pygments==2.19.1
+Pygments==2.19.2
 pymatgen==2025.1.9
 pyparsing==3.2.1
 pyro-api==0.1.2
