### 2025-09-28 rahulkhorana/PolyatomicComplexes#167 @ritesh001
> [@ritesh001](https://github.com/ritesh001) That's a bit odd. The other reviewer could run no issue. Can you tell me what instance you were running on + Python notebook kernel so I can try to reproduce + debug?
>
> Edit: oh you mean there's no cells after the notebook dataset load. Yeah that was intentional. I wanted to show the datasets api yielded the data frame. Not trying to do anything with the data.

Yes, that's what I meant! Everything runs well without any issues in the notebook.
If you do not want to do anything with the data that's fine.
