### 2025-09-29 rahulkhorana/PolyatomicComplexes#167 @rahulkhorana
@RMeli I wanted your take given the current status of this issue. Do you believe we need a minimal working example or just showcasing the API is sufficient. Please let me know. Thanks!
