### 2025-09-29 c30b49d
issue #163 -> informative path message
