+++ polyatomic_complexes/src/complexes/polyatomic_complex.py
@@ -16,6 +16,7 @@ if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
 print("Loading [polyatomic-complexes]")
+print(f"Prepended `{src_dir}` to PATH")
 
 from polyatomic_complexes.src.complexes.core_utils import GluingMap, ElectronField
 from polyatomic_complexes.src.complexes.building_blocks import Electron
