+++ polyatomic_complexes/src/complexes/space_utils.py
@@ -17,6 +17,7 @@ src_dir = project_root
 if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 print("Loading [polyatomic-complexes]")
+print(f"Prepended `{src_dir}` to PATH")
 
 from polyatomic_complexes.src.complexes.building_blocks import Neutron, Proton, Electron
 
