+++ polyatomic_complexes/src/complexes/quantum_complex.py
@@ -35,6 +35,7 @@ if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
 print("Loading [polyatomic-complexes]")
+print(f"Prepended `{src_dir}` to PATH")
 
 from polyatomic_complexes.src.complexes.abstract_complex import AbstractComplex
 
