### 2025-09-29 1641cd5
issue #163 -> remove wildcard match
