+++ polyatomic_complexes/src/complexes/build_atom_lookup.py
@@ -13,6 +13,7 @@ if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
 print("Loading [polyatomic-complexes]")
+print(f"Prepended `{src_dir}` to PATH")
 
 from polyatomic_complexes.src.complexes.atomic_complex import AtomComplex
 
