### 2025-09-29 8bf2423
Update dependabot.yml
