### 2025-09-29 38ca488
Delete .github/dependabot.yml
