+++ .github/dependabot.yml
@@ -1,1 +1,0 @@
-# eliminated
