### 2025-09-29 2323a2f
adding more tests for ci file in ref to issue #166
