+++ polyatomic_complexes/tests/test_geom.py
@@ -338,122 +338,108 @@ def check_force_complex_test(complex: ForceComplex):
 
 
 def check_quantum_complex_test(complex: QuantumComplex):
-    try:
-        forces = complex.forces()
-        assert isinstance(forces, np.ndarray), "Forces should be a numpy.ndarray"
-        assert forces.size != 0, "Forces array should not be empty"
-        assert forces.ndim == 2, "Forces array should be 2D"
-        assert forces.shape[1] == 3, "Forces array should have shape (N, 3)"
-        assert np.any(forces != 0), "Forces should have at least one non-zero component"
-        electrostatics = complex.electrostatic_potentials()
-        assert isinstance(
-            electrostatics, dict
-        ), "Electrostatic potentials should be a dict"
-        grid_coords = electrostatics.get("grid_coords")
-        total_esp = electrostatics.get("total_electrostatic_potential")
-        assert isinstance(
-            grid_coords, list
-        ), "Electrostatic 'grid_coords' should be a list"
-        assert isinstance(
-            total_esp, list
-        ), "Electrostatic 'total_electrostatic_potential' should be a list"
-        assert len(grid_coords) > 0, "Electrostatic 'grid_coords' should not be empty"
+    forces = np.asarray(complex.forces())
+    assert isinstance(forces, np.ndarray), "Forces should be a numpy.ndarray"
+    assert forces.size != 0, "Forces array should not be empty"
+    assert forces.ndim == 2, "Forces array should be 2D"
+    assert forces.shape[1] == 3, "Forces array should have shape (N, 3)"
+    assert np.any(forces != 0), "Forces should have at least one non-zero component"
+    electrostatics = complex.electrostatic_potentials()
+    assert isinstance(electrostatics, dict), "Electrostatic potentials should be a dict"
+    grid_coords = electrostatics.get("grid_coords")
+    total_esp = electrostatics.get("total_electrostatic_potential")
+    assert isinstance(grid_coords, list), "Electrostatic 'grid_coords' should be a list"
+    assert isinstance(
+        total_esp, list
+    ), "Electrostatic 'total_electrostatic_potential' should be a list"
+    assert len(grid_coords) > 0, "Electrostatic 'grid_coords' should not be empty"
+    assert (
+        len(total_esp) > 0
+    ), "Electrostatic 'total_electrostatic_potential' should not be empty"
+    positions = np.asarray(complex.positions())
+    assert isinstance(positions, np.ndarray), "Positions should be a numpy.ndarray"
+    assert positions.size != 0, "Positions array should not be empty"
+    assert positions.ndim == 2, "Positions array should be 2D"
+    assert positions.shape[1] == 3, "Positions array should have shape (N, 3)"
+    dist = np.asarray(complex.distance_matrix())
+    assert isinstance(dist, np.ndarray), "Distance matrix should be a numpy.ndarray"
+    assert dist.size != 0, "Distance matrix should not be empty"
+    assert dist.ndim == 2, "Distance matrix should be 2D"
+    assert dist.shape[0] == dist.shape[1], "Distance matrix should be square"
+    assert np.all(dist >= 0), "All distances should be non-negative"
+    assert not np.all(dist == 0), "Not all distances should be zero"
+    assert len(np.unique(dist)) > 1, "Distance matrix should have distinct values"
+    fermi_level = complex.fermi_level()
+    assert (
+        isinstance(fermi_level, float) or fermi_level is None
+    ), "Fermi level should be a float or None"
+    if fermi_level is not None:
+        assert fermi_level != 0.0, "Fermi level should not be zero"
+    eigenvalues = np.asarray(complex.eigenvalues())
[…]
