### 2025-09-29 a51636b
memory smaller
