### 2025-09-29 45d0c07
attempt reduce mem
