### 2025-10-07 #8828 @RMeli
@rahulkhorana apologies for not noticing before, could you please clarify the following:

1. Has the preprint https://doi.org/10.48550/arXiv.2409.15600 been published anywhere? (Journal, conference proceedings, ...).
2. In the preprint https://doi.org/10.48550/arXiv.2409.15600, which links to the code submitted here in the abstract, there are two other authors. Could you please clarify why the authors in that preprint now appear only in the acknowledgments?
