### 2025-10-07 6b70de7
Addresses Issue #199
