### 2025-10-09 9e9d658
Update paper.md
factoring in @RMeli changes
