### 2025-10-16 #8828 @rahulkhorana
Hi @RMeli,

I just wanted to quickly check in and ask about the overall timeline with remaining events.

Please let me know.

Thank you for your time and consideration in this process.

Best regards,
Rahul
