### 2025-10-18 #8828 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
