### 2025-10-21 #8828 @ritesh001
Congratulations @rahulkhorana on your paper!
